+++ franka_example_controllers/include/franka_example_controllers/joint_position_example_controller.hpp
@@ -16,19 +16,19 @@
 
 #include <string>
 
+#include <Eigen/Eigen>
 #include <controller_interface/controller_interface.hpp>
 #include <rclcpp/rclcpp.hpp>
-
-#include <franka_semantic_components/franka_cartesian_velocity_interface.hpp>
+#include "franka_semantic_components/franka_robot_state.hpp"
 
 using CallbackReturn = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
 
 namespace franka_example_controllers {
 
 /**
- * The elbow example controller
+ * The joint position example controller moves in a periodic movement.
  */
-class ElbowExampleController : public controller_interface::ControllerInterface {
+class JointPositionExampleController : public controller_interface::ControllerInterface {
  public:
   [[nodiscard]] controller_interface::InterfaceConfiguration command_interface_configuration()
       const override;
@@ -39,18 +39,17 @@ class ElbowExampleController : public controller_interface::ControllerInterface 
   CallbackReturn on_init() override;
   CallbackReturn on_configure(const rclcpp_lifecycle::State& previous_state) override;
   CallbackReturn on_activate(const rclcpp_lifecycle::State& previous_state) override;
-  CallbackReturn on_deactivate(const rclcpp_lifecycle::State& previous_state) override;
 
  private:
-  std::unique_ptr<franka_semantic_components::FrankaCartesianVelocityInterface>
-      franka_cartesian_velocity_;
+  std::string arm_id_;
+  const int num_joints = 7;
+  std::array<double, 7> initial_q_{0, 0, 0, 0, 0, 0, 0};
+  const double trajectory_period{0.001};
+  double elapsed_time_ = 0.0;
+  std::string arm_id{"panda"};
 
-  const bool k_elbow_activated_{true};
-  std::vector<double> initial_cartesian_velocity_and_elbow;
   bool initialization_flag_{true};
-  std::array<double, 2> initial_elbow_configuration_{0.0, 0.0};
-  double elapsed_time_{0.0};
-  const double traj_frequency_{0.001};
+  rclcpp::Time start_time_;
 };
 
 }  // namespace franka_example_controllers