+++ franka_example_controllers/src/joint_impedance_example_controller.cpp
@@ -47,11 +47,12 @@ JointImpedanceExampleController::state_interface_configuration() const {
 
 controller_interface::return_type JointImpedanceExampleController::update(
     const rclcpp::Time& /*time*/,
-    const rclcpp::Duration& /*period*/) {
+    const rclcpp::Duration& period) {
   updateJointStates();
   Vector7d q_goal = initial_q_;
-  auto time = this->get_node()->now() - start_time_;
-  double delta_angle = M_PI / 8.0 * (1 - std::cos(M_PI / 2.5 * time.seconds()));
+  elapsed_time_ = elapsed_time_ + period.seconds();
+
+  double delta_angle = M_PI / 8.0 * (1 - std::cos(M_PI / 2.5 * elapsed_time_));
   q_goal(3) += delta_angle;
   q_goal(4) += delta_angle;
 
@@ -111,8 +112,10 @@ CallbackReturn JointImpedanceExampleController::on_configure(
 CallbackReturn JointImpedanceExampleController::on_activate(
     const rclcpp_lifecycle::State& /*previous_state*/) {
   updateJointStates();
+  dq_filtered_.setZero();
   initial_q_ = q_;
-  start_time_ = this->get_node()->now();
+  elapsed_time_ = 0.0;
+
   return CallbackReturn::SUCCESS;
 }