+++ franka_example_controllers/src/joint_position_example_controller.cpp
@@ -1,0 +1,95 @@
+// Copyright (c) 2023 Franka Emika GmbH
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <franka_example_controllers/joint_position_example_controller.hpp>
+
+#include <cassert>
+#include <cmath>
+#include <exception>
+#include <string>
+
+#include <Eigen/Eigen>
+
+namespace franka_example_controllers {
+
+controller_interface::InterfaceConfiguration
+JointPositionExampleController::command_interface_configuration() const {
+  controller_interface::InterfaceConfiguration config;
+  config.type = controller_interface::interface_configuration_type::INDIVIDUAL;
+  for (int i = 1; i <= num_joints; ++i) {
+    config.names.push_back(arm_id_ + "_joint" + std::to_string(i) + "/position");
+  }
+  return config;
+}
+
+controller_interface::InterfaceConfiguration
+JointPositionExampleController::state_interface_configuration() const {
+  return controller_interface::InterfaceConfiguration{
+      controller_interface::interface_configuration_type::NONE};
+}
+
+controller_interface::return_type JointPositionExampleController::update(
+    const rclcpp::Time& /*time*/,
+    const rclcpp::Duration& /*period*/) {
+  if (initialization_flag_) {
+    for (size_t i = 0; i < 7; ++i) {
+      initial_q_.at(i) = command_interfaces_[i].get_value();
+    }
+    initialization_flag_ = false;
+  }
+
+  elapsed_time_ = elapsed_time_ + trajectory_period;
+  double delta_angle = M_PI / 16 * (1 - std::cos(M_PI / 5.0 * elapsed_time_)) * 0.2;
+
+  for (size_t i = 0; i < 7; ++i) {
+    if (i == 4) {
+      command_interfaces_[i].set_value(initial_q_.at(i) - delta_angle);
+    } else {
+      command_interfaces_[i].set_value(initial_q_.at(i) + delta_angle);
+    }
+  }
+
+  return controller_interface::return_type::OK;
+}
+
+CallbackReturn JointPositionExampleController::on_init() {
+  try {
+    auto_declare<std::string>("arm_id", "panda");
+
+  } catch (const std::exception& e) {
+    fprintf(stderr, "Exception thrown during init stage with message: %s \n", e.what());
+    return CallbackReturn::ERROR;
+  }
+  return CallbackReturn::SUCCESS;
+}
+
+CallbackReturn JointPositionExampleController::on_configure(
+    const rclcpp_lifecycle::State& /*previous_state*/) {
+  arm_id_ = get_node()->get_parameter("arm_id").as_string();
+
+  return CallbackReturn::SUCCESS;
+}
+
+CallbackReturn JointPositionExampleController::on_activate(
+    const rclcpp_lifecycle::State& /*previous_state*/) {
+  initialization_flag_ = true;
+  elapsed_time_ = 0.0;
+  return CallbackReturn::SUCCESS;
+}
+
+}  // namespace franka_example_controllers
+#include "pluginlib/class_list_macros.hpp"
+// NOLINTNEXTLINE
+PLUGINLIB_EXPORT_CLASS(franka_example_controllers::JointPositionExampleController,
+                       controller_interface::ControllerInterface)