+++ franka_hardware/src/franka_hardware_interface.cpp
@@ -42,6 +42,7 @@ FrankaHardwareInterface::FrankaHardwareInterface()
     : command_interfaces_info_({
           {hardware_interface::HW_IF_EFFORT, kNumberOfJoints, effort_interface_claimed_},
           {hardware_interface::HW_IF_VELOCITY, kNumberOfJoints, velocity_joint_interface_claimed_},
+          {hardware_interface::HW_IF_POSITION, kNumberOfJoints, position_joint_interface_claimed_},
           {k_HW_IF_ELBOW_COMMAND, hw_elbow_command_names_.size(), elbow_command_interface_claimed_},
           {k_HW_IF_CARTESIAN_VELOCITY, hw_cartesian_velocities_.size(),
            velocity_cartesian_interface_claimed_},
@@ -74,9 +75,11 @@ std::vector<CommandInterface> FrankaHardwareInterface::export_command_interfaces
   command_interfaces.reserve(info_.joints.size());
   for (auto i = 0U; i < info_.joints.size(); i++) {
     command_interfaces.emplace_back(CommandInterface(
-        info_.joints[i].name, hardware_interface::HW_IF_EFFORT, &hw_commands_.at(i)));
+        info_.joints[i].name, hardware_interface::HW_IF_EFFORT, &hw_effort_commands_.at(i)));
     command_interfaces.emplace_back(CommandInterface(
-        info_.joints[i].name, hardware_interface::HW_IF_VELOCITY, &hw_commands_.at(i)));
+        info_.joints[i].name, hardware_interface::HW_IF_VELOCITY, &hw_velocity_commands_.at(i)));
+    command_interfaces.emplace_back(CommandInterface(
+        info_.joints[i].name, hardware_interface::HW_IF_POSITION, &hw_position_commands_.at(i)));
   }
 
   // cartesian velocity command interface 6 in order: dx, dy, dz, wx, wy, wz
@@ -97,7 +100,6 @@ std::vector<CommandInterface> FrankaHardwareInterface::export_command_interfaces
 
 CallbackReturn FrankaHardwareInterface::on_activate(
     const rclcpp_lifecycle::State& /*previous_state*/) {
-  hw_commands_.fill(0);
   read(rclcpp::Time(0),
        rclcpp::Duration(0, 0));  // makes sure that the robot state is properly initialized.
   RCLCPP_INFO(getLogger(), "Started");
@@ -112,6 +114,24 @@ CallbackReturn FrankaHardwareInterface::on_deactivate(
   return CallbackReturn::SUCCESS;
 }
 
+template <typename CommandType>
+void initializeCommand(bool& first_update,
+                       const bool& interface_running,
+                       CommandType& hw_command,
+                       const CommandType& new_command) {
+  if (first_update && interface_running) {
+    hw_command = new_command;
+    first_update = false;
+  }
+}
+
+void FrankaHardwareInterface::initializePositionCommands(const franka::RobotState& robot_state) {
+  initializeCommand(first_elbow_update_, elbow_command_interface_running_, hw_elbow_command_,
+                    robot_state.elbow_c);
+  initializeCommand(first_position_update_, position_joint_interface_running_,
+                    hw_position_commands_, robot_state.q_d);
+}
+
 hardware_interface::return_type FrankaHardwareInterface::read(const rclcpp::Time& /*time*/,
                                                               const rclcpp::Duration& /*period*/) {
   if (hw_franka_model_ptr_ == nullptr) {
@@ -120,11 +140,7 @@ hardware_interface::return_type FrankaHardwareInterface::read(const rclcpp::Time
 
   hw_franka_robot_state_ = robot_->readOnce();
 
-  if (first_pass_ && elbow_command_interface_running_) {
-    // elbow command should be initialized with the first elbow_c read from the robot
-    hw_elbow_command_ = hw_franka_robot_state_.elbow_c;
-    first_pass_ = false;
-  }
+  initializePositionCommands(hw_franka_robot_state_);
 
   hw_positions_ = hw_franka_robot_state_.q;
   hw_velocities_ = hw_franka_robot_state_.dq;
@@ -133,23 +149,28 @@ hardware_interface::return_type FrankaHardwareInterface::read(const rclcpp::Time
   return hardware_interface::return_type::OK;
 }
 
+template <typename CommandType>
+bool hasInfinite(const CommandType& commands) {
+  return std::any_of(commands.begin(), commands.end(),
+                     [](double command) { return !std::isfinite(command); });
+}
+
 hardware_interface::return_type FrankaHardwareInterface::write(const rclcpp::Time& /*time*/,
                                                                const rclcpp::Duration& /*period*/) {
-  if (std::any_of(hw_commands_.begin(), hw_commands_.end(),
-                  [](double hw_command) { return !std::isfinite(hw_command); })) {
+  if (hasInfinite(hw_position_commands_) || hasInfinite(hw_effort_commands_) ||
+      hasInfinite(hw_velocity_commands_) || hasInfinite(hw_cartesian_velocities_) ||
+      hasInfinite(hw_elbow_command_)) {
     return hardware_interface::return_type::ERROR;
   }
 
-  if (std::any_of(
-          hw_cartesian_velocities_.begin(), hw_cartesian_velocities_.end(),
-          [](double hw_cartesian_velocity) { return !std::isfinite(hw_cartesian_velocity); })) {
-    return hardware_interface::return_type::ERROR;
-  }
-
-  if (velocity_joint_interface_running_ || effort_interface_running_) {
-    robot_->writeOnce(hw_commands_);
+  if (velocity_joint_interface_running_) {
+    robot_->writeOnce(hw_velocity_commands_);
+  } else if (effort_interface_running_) {
+    robot_->writeOnce(hw_effort_commands_);
+  } else if (position_joint_interface_running_ && !first_position_update_) {
+    robot_->writeOnce(hw_position_commands_);
   } else if (velocity_cartesian_interface_running_ && elbow_command_interface_running_ &&
-             !first_pass_) {
+             !first_elbow_update_) {
     // Wait until the first read pass after robot controller is activated to write the elbow
     // command to the robot
     robot_->writeOnce(hw_cartesian_velocities_, hw_elbow_command_);
@@ -171,13 +192,14 @@ CallbackReturn FrankaHardwareInterface::on_init(const hardware_interface::Hardwa
   }
 
   for (const auto& joint : info_.joints) {
-    if (joint.command_interfaces.size() != 2) {
-      RCLCPP_FATAL(getLogger(), "Joint '%s' has %zu command interfaces found. 2 expected.",
+    if (joint.command_interfaces.size() != 3) {
+      RCLCPP_FATAL(getLogger(), "Joint '%s' has %zu command interfaces found. 3 expected.",
                    joint.name.c_str(), joint.command_interfaces.size());
       return CallbackReturn::ERROR;
     }
     if (joint.command_interfaces[0].name != hardware_interface::HW_IF_EFFORT &&
-        joint.command_interfaces[0].name != hardware_interface::HW_IF_VELOCITY) {
+        joint.command_interfaces[0].name != hardware_interface::HW_IF_VELOCITY &&
+        joint.command_interfaces[0].name != hardware_interface::HW_IF_POSITION) {
       RCLCPP_FATAL(getLogger(),
                    "Joint '%s' has unexpected command interface '%s'. Expected '%s' and '%s' ",
                    joint.name.c_str(), joint.command_interfaces[0].name.c_str(),
@@ -238,7 +260,7 @@ hardware_interface::return_type FrankaHardwareInterface::perform_command_mode_sw
     const std::vector<std::string>& /*start_interfaces*/,
     const std::vector<std::string>& /*stop_interfaces*/) {
   if (!effort_interface_running_ && effort_interface_claimed_) {
-    hw_commands_.fill(0);
+    hw_effort_commands_.fill(0);
     robot_->stopRobot();
     robot_->initializeTorqueInterface();
     effort_interface_running_ = true;
@@ -248,7 +270,7 @@ hardware_interface::return_type FrankaHardwareInterface::perform_command_mode_sw
   }
 
   if (!velocity_joint_interface_running_ && velocity_joint_interface_claimed_) {
-    hw_commands_.fill(0);
+    hw_velocity_commands_.fill(0);
     robot_->stopRobot();
     robot_->initializeJointVelocityInterface();
     velocity_joint_interface_running_ = true;
@@ -257,12 +279,23 @@ hardware_interface::return_type FrankaHardwareInterface::perform_command_mode_sw
     velocity_joint_interface_running_ = false;
   }
 
+  if (!position_joint_interface_running_ && position_joint_interface_claimed_) {
+    robot_->stopRobot();
+    robot_->initializeJointPositionInterface();
+    position_joint_interface_running_ = true;
+    first_position_update_ = true;
+  } else if (position_joint_interface_running_ && !position_joint_interface_claimed_) {
+    robot_->stopRobot();
+    position_joint_interface_running_ = false;
+  }
+
   if (!velocity_cartesian_interface_running_ && velocity_cartesian_interface_claimed_) {
     hw_cartesian_velocities_.fill(0);
     robot_->stopRobot();
     robot_->initializeCartesianVelocityInterface();
     if (!elbow_command_interface_running_ && elbow_command_interface_claimed_) {
       elbow_command_interface_running_ = true;
+      first_elbow_update_ = true;
     }
     velocity_cartesian_interface_running_ = true;
   } else if (velocity_cartesian_interface_running_ && !velocity_cartesian_interface_claimed_) {